Move load_html_texts() to textdata.py
