+++ server/server/data_loader/tests/test_textdata.py
@@ -3,7 +3,10 @@ from pathlib import Path
 
 import pytest
 
-from data_loader.textdata import TextInfoModel, should_process_file
+from common.arangodb import get_db
+from common.utils import current_app
+from data_loader.change_tracker import ChangeTracker
+from data_loader.textdata import TextInfoModel, should_process_file, load_html_texts
 
 
 class TextInfoModelSpy(TextInfoModel):
@@ -277,3 +280,117 @@ class TestShouldProcessFile:
         absolute_path.touch()
         files_to_process = {'xyz.html' : 0,}
         assert not should_process_file(base_path, files_to_process, absolute_path)
+
+
+class TestLoadHtmlTexts:
+    @pytest.fixture
+    def database(self):
+        app_ = current_app()
+        app_.config['ARANGO_DB'] = 'suttacentral_data_load_tests'
+
+        with app_.app_context():
+            db = get_db()
+            db.collection('mtimes').truncate()
+            db.collection('html_text').truncate()
+            yield db
+
+    @pytest.fixture
+    def sc_data_dir(self, tmp_path) -> Path:
+        return tmp_path
+
+    @pytest.fixture
+    def html_dir(self, sc_data_dir) -> Path:
+        path = sc_data_dir / 'html_text'
+        path.mkdir()
+        return path
+
+    @pytest.fixture
+    def html(self) -> str:
+        return (
+        "<html>"
+        "<head>"
+        "<meta author='Bhikkhu Bodhi'>"
+        "</head>"
+        "<body>"
+        "<header><h1>1. The Root of All Things</h1></header>"
+        "<span class='publication-date'>2009</span>"
+        "</body>"
+        "</html>"
+    )
+
+    @pytest.fixture
+    def tracker(self, sc_data_dir, database):
+        return ChangeTracker(base_dir=sc_data_dir, db=database)
+
+    @pytest.mark.skip('Long running test.')
+    def test_load_from_repository(self, database):
+        sc_data_dir = Path('/opt/sc/sc-flask/sc-data/')
+        html_dir = Path('/opt/sc/sc-flask/sc-data/html_text')
+        tracker = ChangeTracker(base_dir=sc_data_dir, db=database)
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+
+    def test_load_empty_html_dir(self, tracker, database, sc_data_dir, html_dir):
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+        assert database.collection('html_text').count() == 0
+
+    def test_load_empty_language_dir(self, tracker, database, sc_data_dir, html_dir):
+        language_dir = html_dir / 'en'
+        language_dir.mkdir()
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+        assert database.collection('html_text').count() == 0
+
+    def test_load_one_text(self, database, sc_data_dir, html_dir, html):
+        language_dir = html_dir / 'en'
+        language_dir.mkdir()
+
+        sutta_path = language_dir / 'mn1.html'
+        sutta_path.write_text(html)
+
+        tracker = ChangeTracker(base_dir=sc_data_dir, db=database)
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+
+        assert database.collection('html_text').count() == 1
+
+    def test_skip_unmodified_text(self, database, sc_data_dir, html_dir, html):
+        language_dir = html_dir / 'en'
+        language_dir.mkdir()
+
+        sutta_path = language_dir / 'mn1.html'
+        sutta_path.write_text(html)
+
+        tracker = ChangeTracker(base_dir=sc_data_dir, db=database)
+        tracker.update_mtimes()
+        tracker = ChangeTracker(base_dir=sc_data_dir, db=database)
+
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+
+        assert database.collection('html_text').count() == 0
+
+    def test_skip_files_not_in_language_subdirectories(self, database, sc_data_dir, html_dir, html):
+        skip_path = html_dir / 'skip_me.html'
+        skip_path.write_text(html)
+
+        language_dir = html_dir / 'en'
+        language_dir.mkdir()
+        ok_path = language_dir / 'mn1.html'
+        ok_path.write_text(html)
+
+        tracker = ChangeTracker(base_dir=sc_data_dir, db=database)
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+
+        assert database.collection('html_text').count() == 1
+
+    def test_load_files_in_subdirectories(self, database, sc_data_dir, html_dir, html):
+        language_dir = html_dir / 'en'
+        language_dir.mkdir()
+
+        collection_dir = language_dir / 'mn'
+        collection_dir.mkdir()
+
+        sutta_path = collection_dir / 'mn1.html'
+        sutta_path.write_text(html)
+
+        tracker = ChangeTracker(base_dir=sc_data_dir, db=database)
+        load_html_texts(change_tracker=tracker, db=database, data_dir=sc_data_dir, html_dir=html_dir)
+
+        assert database.collection('html_text').count() == 1
